+++ infobot.lua
@@ -9,6 +9,7 @@ function OnStartup()
 		sChatFile = "chatcore.lua",
 		sReportBot = "#[Hub-Feed]",
 		sHelp = "",
+		sAllCategories = "",
 		iModProfile = 4,
 		iRegProfile = 5
 	}
@@ -35,7 +36,7 @@ function OnStartup()
 		fHelp:close()
 	end
 	Core.RegBot( tCfg.sBotName, tCfg.sBotDescription, tCfg.sBotEmail, true )
-	sAllModerators, sAllCategories = tFunction.Connect()
+	tCfg.sAllCategories = tFunction.Connect()
 end
 
 function ToArrival( tUser, sMessage )
@@ -60,7 +61,7 @@ function ExecuteCommand( tUser, sCommand, sData )
 		else
 			iLimit = ( tonumber(sData) > 35 and 35 ) or tonumber(sData)
 		end
-		local sList = tInfobot.readAll( "all", iLimit )
+		local sList = tInfobot.readAll( iLimit )
 		if sList then
 			Core.SendPmToUser( tUser, tCfg.sBotName, sList )
 		end
@@ -74,6 +75,19 @@ function ExecuteCommand( tUser, sCommand, sData )
 			iLimit = ( tonumber(sData) > 50 and 50 ) or tonumber(sData)
 		end
 		local sList = tInfobot.readOne( "requests", iLimit )
+		if sList then
+			Core.SendPmToUser( tUser, tCfg.sBotName, sList )
+			sList = nil
+		end
+		return true
+
+	elseif sCommand == "readbuysell" or sCommand == "rbsell" then
+		if ( not sData ) or ( sData and not tonumber(sData) ) then
+			iLimit = 15
+		else
+			iLimit = ( tonumber(sData) > 50 and 50 ) or tonumber(sData)
+		end
+		local sList = tInfobot.readOne( "buynsell", iLimit )
 		if sList then
 			Core.SendPmToUser( tUser, tCfg.sBotName, sList )
 			sList = nil
@@ -120,6 +134,11 @@ function ExecuteCommand( tUser, sCommand, sData )
 		return true
 	end
 
+	if sData:len() == 0 then
+		Core.SendPmToUser( tUser, tCfg.sBotName, "You did not pass anything as argument." )
+		return false
+	end
+
 	local tInsertData, iLastID, sAdditionNotify = { sMsg = sData, sTable = "all" }, 0, "New %s has been added to [ %s ] table. Use %s for more information."
 
 	if not tInsertData.sMsg then
@@ -135,7 +154,7 @@ function ExecuteCommand( tUser, sCommand, sData )
 		if tFunction.CheckCategory( tBreak[1] ) then
 			tInsertData.sCtg = tBreak[1]
 		else
-			Core.SendPmToUser( tUser, tCfg.sBotName, "Wrong category!" )
+			Core.SendPmToUser( tUser, tCfg.sBotName, "Wrong category! "..tCfg.sAllCategories )
 			return false
 		end
 		tInsertData.sMsg, tInsertData.sTable = table.concat( tBreak, " ", 2 ), "requests"
@@ -154,9 +173,9 @@ function ExecuteCommand( tUser, sCommand, sData )
 	elseif sCommand == "asug" or sCommand == "asg" then
 		local tBreak = tFunction.Explode( sData )
 		if tFunction.CheckCategory( tBreak[1] ) then
-			tInsertData.sCtg = tBreak[1]
+			tInsertData.sCtg = tBreak[1]:lower()
 		else
-			Core.SendPmToUser( tUser, tCfg.sBotName, "Wrong category!" )
+			Core.SendPmToUser( tUser, tCfg.sBotName, "Wrong category! "..tCfg.sAllCategories )
 			return false
 		end
 		tInsertData.sMsg, tInsertData.sTable = table.concat( tBreak, " ", 2 ), "suggestions"
@@ -214,6 +233,54 @@ function ExecuteCommand( tUser, sCommand, sData )
 		Core.SendPmToUser( tUser, tCfg.sBotName, "Deletion info has been added at ID #"..tostring(iLastID).."." )
 		return true
 
+	elseif sCommand == "addbns" or sCommand == "absell" then
+		local tBreak = tFunction.Explode( sData )
+		local _, sError = tFunction.CheckBnS( string.lower(tBreak[1]) )
+		if _ then
+			tInsertData.sType = _
+		else
+			Core.SendPmToUser( tUser, tCfg.sBotName, "Wrong category! Available ones are \n"..sError )
+			return false
+		end
+		tInsertData.sMsg, tInsertData.sTable = table.concat( tBreak, " ", 2 ), "buynsell"
+		if tInsertData.sMsg:lower():find( "[%[buy|sell%]]" ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "You don't need to separately tag the thread." )
+			return false
+		end
+		iLastID = tInfobot.add( tUser, tInsertData )
+		if ( not tonumber(iLastID) ) or ( tonumber(iLastID) == 0 ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "Something went wrong." )
+			return false
+		end
+		local sChatMessage = sAdditionNotify:format( ("%s: %s"):format(tBreak[1]:upper(), tInsertData.sMsg), tInsertData.sTable:upper(), tCfg.sBotName )
+		SendToRoom( tUser.sNick, sChatMessage, tCfg.sReportBot )
+		tFunction.SendToAll( tUser.sNick, sChatMessage )
+		sChatMessage = nil
+		Core.SendPmToUser( tUser, tCfg.sBotName, "Buy and sell entry has been added at ID #"..tostring(iLastID).."." )
+		return true
+
+	elseif sCommand == "addreply" or sCommand == "amsg" then
+		local tBreak = tFunction.Explode( sData )
+		if not tonumber( tBreak[1], 10 ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "The ID must be numerical!" )
+			return false
+		end
+		tInsertData.sMsg, tInsertData.iID, tInsertData.sTable = table.concat( tBreak, " ", 2 ), tonumber( tBreak[1], 10 ), "replies"
+		local tRow = tFunction.FetchRow( "buynsell", tInsertData.iID )
+		if not tRow then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "No entry with that ID." )
+			return false
+		end
+		iLastID = tInfobot.add( tUser, tInsertData )
+		if ( not tonumber(iLastID) ) or ( tonumber(iLastID) == 0 ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "Something went wrong." )
+			return false
+		end
+		tInfobot.StoreMessage( tUser.sNick, tRow.nick, "I've replied to your buy and sell thread. ID#"..tostring(tInsertData.iID).." - "..tRow.msg.."." )
+		local sReply = ("Your reply to buy and sell thread#%d has been added at ID #%d. %s will be notified with your message. Keep checking the thread for further replies."):format( tInsertData.iID, iLastID, tRow.nick )
+		Core.SendPmToUser( tUser, tCfg.sBotName, sReply )
+		return true
+
 	elseif sCommand == "fill" or sCommand == "freq" then
 		local tBreak, sReply = tFunction.Explode( sData ), "You filled the request \n\tID#%d. [%s] - %s (Added by %s)\n\nThe requesting user will be notified with message ID#%d when they connect."
 		if not tonumber( tBreak[1], 10 ) then
@@ -246,6 +313,36 @@ function ExecuteCommand( tUser, sCommand, sData )
 		tInfobot.fill( tUser, tInsertData.iID, true )
 		Core.SendPmToUser( tUser, tCfg.sBotName, sReply:format(tInsertData.iID, tRow.ctg, tRow.msg, tRow.nick, tonumber(iOfflineMessageID)) )
 		return true
+
+	elseif sCommand == "switch" then
+		local tBreak, sReply = tFunction.Explode( sData ), "You switched the status of thread \n\tID#%d. [%s] - %s (Added by %s)\n\nThe requesting user will be notified with message ID#%d when they connect."
+		if not tonumber( tBreak[1], 10 ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "The ID must be numerical!" )
+			return false
+		end
+		tInsertData.sMsg, tInsertData.iID = nil, tonumber( tBreak[1], 10 )
+		local tRow = tFunction.FetchRow( "buynsell", tInsertData.iID )
+		if not tRow then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "No entry with that ID." )
+			return false
+		end
+		if tRow.type:lower() ~= "buy" or tRow.type:lower() ~= "sell" then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "The thread is of [ "..tRow.type.." ] type. Can not switch its status." )
+			return false
+		end
+		if tUser.sNick:lower() == tRow.nick:lower() or tProfiles.AllowMods[tUser.iProfile] then
+			tInfobot.switch( tInsertData.iID )
+			if tUser.sNick:lower() ~= tRow.nick:lower() then
+				local iOfflineMessageID = tInfobot.StoreMessage( tUser.sNick, tRow.nick, "I've switched your buy and sell thread ID#"..tostring(tInsertData.iID).." - "..tRow.msg.."." )
+				Core.SendPmToUser( tUser, tCfg.sBotName, sReply:format(tInsertData.iID, tRow.type, tRow.msg, tRow.nick, tonumber(iOfflineMessageID)) )
+			else
+				Core.SendPmToUser( tUser, tCfg.sBotName, "Status switched successfully." )
+			end
+			return true
+		else
+			Core.SendPmToUser( tUser, tCfg.sBotName, "You do not have sufficient privileges or you didn't start the thread." )
+			return false
+		end
 
 	elseif sCommand == "delreq" or sCommand == "dr" then
 		local tBreak = tFunction.Explode( sData )
@@ -308,6 +405,48 @@ function ExecuteCommand( tUser, sCommand, sData )
 		else
 			Core.SendPmToUser( tUser, tCfg.sBotName, "You are not allowed to delete that news content." )
 			return false
+		end
+		return true
+
+	elseif sCommand == "delbns" or sCommand == "dbsell" then
+		local tBreak = tFunction.Explode( sData )
+		if not tonumber( tBreak[1], 10 ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "The ID must be numerical!" )
+			return false
+		end
+		tInsertData.sMsg, tInsertData.iID, tInsertData.sTable = nil, tonumber( tBreak[1], 10 ), "buynsell"
+		local tRow = tFunction.FetchRow( tInsertData.sTable, tInsertData.iID )
+		if not tRow then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "No entry with that ID." )
+			return false
+		end
+		if tProfiles.AllowVIP[tUser.iProfile] or tRow.nick:lower() == tUser.sNick:lower() then
+			local sTemporary = "The following entry has been deleted:\n\tID#%d. [%s] - %s (Added by %s)"
+			Core.SendPmToUser( tUser, tCfg.sBotName, sTemporary:format(tInsertData.iID, tRow.type, tRow.msg, tRow.nick) )
+			tInfobot.del( tUser, tInsertData )
+		else
+			Core.SendPmToUser( tUser, tCfg.sBotName, "You are not allowed to delete that thread." )
+		end
+		return true
+
+	elseif sCommand == "delmsg" or sCommand == "dmsg" then
+		local tBreak = tFunction.Explode( sData )
+		if not tonumber( tBreak[1], 10 ) then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "The ID must be numerical!" )
+			return false
+		end
+		tInsertData.sMsg, tInsertData.iID, tInsertData.sTable = nil, tonumber( tBreak[1], 10 ), "replies"
+		local tRow = tFunction.FetchRow( tInsertData.sTable, tInsertData.iID )
+		if not tRow then
+			Core.SendPmToUser( tUser, tCfg.sBotName, "No reply with that ID." )
+			return false
+		end
+		if tProfiles.AllowVIP[tUser.iProfile] or tRow.nick:lower() == tUser.sNick:lower() then
+			local sTemporary = "The following reply has been deleted:\n\tID#%d. %s (Added by %s)"
+			Core.SendPmToUser( tUser, tCfg.sBotName, sTemporary:format(tInsertData.iID, tRow.msg, tRow.nick) )
+			tInfobot.del( tUser, tInsertData )
+		else
+			Core.SendPmToUser( tUser, tCfg.sBotName, "You are not allowed to delete that reply." )
 		end
 		return true