+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+2013-04-09
+----------
+
+- Buy and sell implemented.
+- Minor tweaks as listed below:
+    1. Minimum length of message must be 20.
+    2. Categories passed will be lowercase.
+    3. Previously `readall` command repeated a particular message for all subsequent tables.
+- Help file updated with new commands.
+- Changed `INNER JOIN` syntax for `dl` command to `LEFT JOIN` to allow deletions for entries without magnets. Using the same to delete replies for Buy and Sell.
+- Original users will be notified of any mesages/replies on their buy and sell thread.
+
 2013-04-07
 ----------